+++ jdaviz/core/template_mixin.py
@@ -27,8 +27,10 @@ from ipywidgets import widget_serialization
 from ipypopout import PopoutButton
 
 from jdaviz import __version__
-from jdaviz.core.events import (AddDataMessage, RemoveDataMessage,
-                                ViewerAddedMessage, ViewerRemovedMessage)
+from jdaviz.core.events import (
+    AddDataMessage, RemoveDataMessage, ViewerAddedMessage, ViewerRemovedMessage,
+    ViewerRenamedMessage
+)
 from jdaviz.core.user_api import UserApiWrapper, PluginUserApi
 from jdaviz.utils import get_subset_type
 
@@ -1324,6 +1326,7 @@ class ViewerSelect(SelectPluginComponent):
 
         self.hub.subscribe(self, ViewerAddedMessage, handler=self._on_viewers_changed)
         self.hub.subscribe(self, ViewerRemovedMessage, handler=self._on_viewers_changed)
+        self.hub.subscribe(self, ViewerRenamedMessage, handler=self._on_viewers_changed)
 
         # initialize viewer_items from original viewers
         self._on_viewers_changed()