+++ CHANGES.rst
@@ -60,6 +60,8 @@ Specviz2d
 Other Changes and Additions
 ---------------------------
 
+- Add support for changing viewer reference names [#2338]
+
 3.6.2 (unreleased)
 ==================